KokkosKernels,Tpetra,Xpetra,MueLu: Detemplatized Kokkos CRS row view
@trilinos/tpetra @trilinos/xpetra @trilinos/muelu

KokkosSparse::CrsMatrix has a way to get a "view" of a row.  The
methods for getting a row (row(i) and rowConst(i)) and the types of
the views themselves (SparseRowView resp. SparseRowViewConst) took
"SizeType" as a template parameter.  Per #417, this is not necessary,
because the number of entries in a row must always fit in ordinal_type
(as long as the row has no duplicate entries).  Even if the row has a
stride > 1 (which it can only have with ELLPACK or JAG format),
SparseRowView* can figure out the correct type of its internal index
from the MatrixType template parameter.

Thus, I got rid of the SizeType template parameter in both the row(i)
and rowConst(i) methods, and the SparseRowView and SparseRowViewConst
structs.  This should help reduce build time, without reducing
generality.

I had to change Xpetra and MueLu, since they depend on these methods
and structs.  Note that the preferred way to use row(i) and
rowConst(i) is to let the compiler deduce the return type via 'auto'.
Don't actually specify the return type.  This gives us flexibility for
compile-time polymorphism on the return type.  For example, we might
like to get rid of the 'stride' in SparseRowView* -- if users use
'auto' for the row view type, then we could have different view types
for different matrix types, without actually changing users' code.