First prototype of the new ParameterList validation loading method.
The basic idea was to change the way parameters were defined.
Before they all came from a giant XML format string.
Now they can be defined in individual classes, algorithms, and soon - models as well.
So pulp parameters for example don't exist except when using that system.

This is a prototype and generally working but there are several technical issues to resolve.
A few unit tests rely on a relic version which still loads the old XML format.
These will be refactored to eliminate that relic function soon.