Improvements to the new ParameterList method.
This fixes some of the unit tests to not rely on the old Relic form of parameter list validation.
We still have one case in EvaluatePartition which needs further consideration.

Also made some improvements to the handling of the algorithm parameter.
Now we pass the input parameters while loading the validation parameters.
Those can be referenced to determine what parametes are needed.
This avoids the need to store the algorithm name ahead of time.
I'm anticipating some of these methods may become statics so this will smooth that process.