Tpetra: Fixed Distributor createfromsendsandrecvs test w/ non-MPI (#488)
@trilinos/tpetra I fixed Distributor's new createfromsendsandrecvs
test for non-MPI builds (it was already passing for MPI builds).  It
turns out the test was assuming that a fill-complete CrsGraph always
has an Import object.  This is definitely NOT the case when the
CrsGraph's communicator only has one process, as in (for example) a
non-MPI build.

I changed the test so that it only builds and runs in an MPI build.
Furthermore, in case somebody changes the CMakeLists.txt file
mistakenly in the future, the test calls itself "passed" and returns
early if the Import object is null and the communicator has only 1
process (as in, for example, a non-MPI build).

The test now passes.