Tpetra::CrsGraph::getLocalKokkosRowView: Fix thread safety issue
@trilinos/tpetra Tpetra::CrsGraph::getLocalKokkosRowView returns an
unmanaged Kokkos::View of the local column indices of a given row.
This method must be thread safe.  However, it was creating a
Teuchos::ArrayView which views an entry of lclInds1D_
(Teuchos::ArrayRCP<Teuchos::Array<LocalOrdinal> >).  This may not be
thread safe in a debug build, because it may modify an ArrayRCP
reference count.  This commit avoids creating a Teuchos::ArrayView
instance, thus circumventing the problem.