Moved locally defined elementwise operation functors to external namespace to accomodate a compiler problem involving CUDA
error: A type that is defined inside a class and has private or protected access ("ROL::TpetraMultiVector<RealT, LO, GO, Node>::unaryFunc") cannot be used in the template argument type of a __global__ function template instantiation, unless the class is local to a __device__ or __global__ function

Compiling for device causes anger (I.e. build error) for local functors inside a class that are private/protected.

Thanks to Nathan Ellingwood for discovering this error.