TeuchosKokkosCompat,Tpetra: Fix putenv-related build warning
@trilinos/tpetra Kokkos::Compat::KokkosDeviceWrapperNode has been
reporting a build warning only in CUDA builds: "conversion from a
string literal to "char *" is deprecated".  This comes from a call to
putenv that sets the environment variable CUDA_LAUNCH_BLOCKING=1 (as
Tpetra currently requires, due to its use of UVM).  putenv takes its
input as a nonconst pointer.  This causes the build warning when using
a string literal, in this case "CUDA_LAUNCH_BLOCKING=1", as the input
argument.

It turns out that this call isn't really safe.  This is why POSIX
prefers use of setenv over putenv:

http://pubs.opengroup.org/onlinepubs/009695399/functions/putenv.html

In particular, putenv does not let us fix the above build warning by
copying into a temporary nonconst char array and passing that into
putenv.  This is because putenv is free to keep the original input
pointer, rather than copying the input string.  Thus, if anyone tries
to read the environment variable (via getenv) after the temporary
array has been freed, they would be reading invalid memory.  The above
POSIX standard entry alludes to this: "A potential error is to call
putenv() with an automatic variable as the argument, then return from
the calling function while string is still part of the environment."

This commit fixes the warning by using setenv instead of putenv.  It
also does error checking, using setenv's return code and errno.