This change was necessary to have the Catalyst-Sierra plug in work properly with ParaView 5 (and greater)
The problem was that in the dlopen() call in the plugin loadin in the file
/scratch/jamauld/sierra2/TPLs_src/Trilinos/packages/stk/stk_classic/stk_util/stk_util/diag/UserPlugin.cpp
the flag RTDL_GLOBAL was not supplied, so the symbols were local.

In ParaView 4.x this was okay, but in ParaView 5.x paraview uses GLEW
(opengl extensions wrangler) to obtain symbols, and glew does this by making a dlopen() call with
no arguments--and that later dlopen() call in glew won't see the symbols if the dlopen() call in
Sierra doesn't have the RTDL_GLOBAL flag supplied.

Rather than trying to alter trilinos, we fix the problem by calling dlopen() again in
Iovs_DatabaseIO.C to promote the symbols to global with the RTDL_GLOBAL flag.  This is
a standard use of dlopen() and is mentioned in the dlopen man page.  By putting the
fix here we do not need to affect other plugins.

The one detriment is that we can no longer use the command 'load user
plugin file' to preload a different version of the catalyst-sierra plugin than
the system-installed version, as Iovs_DatabaseIO.C doesn't have access to
a filepath specified that way in the input deck (a feature that could be added
to the plugin management if we need to).  However, you can still specify a
non-default sierra-catalyst plugin .so file using the environment variable
CATALYST_PLUGIN, so the capability of using a different-than-standard plugin
file is still avaialble.

(From STK/SEACAS)