Stratimikos: Add #535 test (it FAILS)
@trilinos/stratimikos @trilinos/thyra @trilinos/ifpack2

I added a test for #535.  @nschloe wrote it and I modified it to add
diagnostics.  The test FAILS for me.  I'm pushing it so that we can
see the failing test on the Dashboard.  This requires forcing the
push.

In order to build this test, I had to add Ifpack2 and
ThyraTpetraAdapters as optional test dependencies of Stratimikos.