MueLu: add new unit test for blocked rebalancing
This unit test covers RebalanceBlockAcFactory as well as the IsorropiaInterface
(if Epetra is enabled and MueLu has been compiled with Isorropia). For Tpetra
(or Epetra without Isorropia enabled) we build a dummy partition on two processors.

The unit test should run on 3 or more processors. Otherwise it will not test anything
reasonable.