Fix stk ngp unit test. I broke the jamfile a few days ago and it wasn't building, until now...
Also, fixed template argument to ngp::Mesh's connectivity View.  It was not templated on the MemSpace.
So when a MemSpace templated View was being copying into the non-MemSpace tempated View, most of the time
this worked without error.  BUT, when UVM is on, the default memory space is CudaUVM, while MemSpace is
specified in stk to be Cuda.  This is an invalid assignment.  (Alternatively, stk mesh could potentially
just choose to get rid of the MemSpace template argument, and just let Kokkos auto-decide the device
memory space.)  This fix will allow ngp::Mesh to be used from adagio, where UVM is currently turned on
for the existing capabilities.

(From STK/SEACAS)