Ifpack2: reenable AdditiveSchwarz test
The AdditiveSchwarz test w/ sparse direct solves on subdomains was not
running.  This commit reenables the test.  The issue was that
Amesos2_config.h was not being included, so the preprocessor macros
guarding the test were not properly defined.

I've also renamed the test and generalized it so that either SuperLU or
KLU2 can be used.

There is still a problem, in that for overlap>0, the test will fail.  I
am still investigating this in git issue #558.

Build/Test Cases Summary
Enabled Packages: Ifpack2
Enabled all Forward Packages
0) MPI_DEBUG => passed: passed=29,notpassed=0 (12.27 min)
1) SERIAL_RELEASE => passed: passed=25,notpassed=0 (7.54 min)