MueLu: additive Hierarchy::Iterate supported
MueLu_Hierarchy_def.hpp is modified to have an additive Two_Level
Schwarz preconditioner along with the standard recursive multiplicative
version.
Cuurently there is a cmake configure variable to allow to switch
from one version to the other, requiring recompiling the code.
It may be better to change the policy so that the choice of
Hierarchy::Iterate can be made at runtime.