Tpetra: added check on #rows in exception
An extra condition has been added in the exception checking that the LU
factors be actually lower triangular and upper triangular. This was
needed because the exception is thrown when one processor does not own
any row in the RowMap associated with a Tpetra::CrsMatrix object.  Now
the exception also checks that #rows on the node is strictly positive.

Build/Test Cases Summary
Enabled Packages: TpetraCore, TpetraKernels
Disabled Packages: ROL
Enabled all Forward Packages
0) MPI_DEBUG => passed: passed=467,notpassed=0 (13.93 min)
1) SERIAL_RELEASE => passed: passed=432,notpassed=0 (7.72 min)
2) MPI_SS_DEBUG => passed: passed=743,notpassed=0 (14.55 min)
3) SERIAL_SS_RELEASE => passed: passed=443,notpassed=0 (19.38 min)
Other local commits for this build/test group: c470682