Panzer: Handle (side ID a, side ID b) complexity in interface condition.
Meshes can be created such that side IDs along a sideset are different. Panzer
already handles this in its BC machinery; Panzer makes one field manager per
side ID in a BC.

An interface condition adds a layer of complexity. Consider an interface with
elements having side IDs A and B, each a list. Now we need to break the elements
up not just by unique side ID in each separately, but rather by unique (side ID
a, side ID b) pair. This commit does that.

In addition, add a test to the PoissonInterfaceExample test set that exercises
this new capability. (Thanks to S. Bond for generating an interesting mesh for
this test.)