Zoltan2: Parameter list updates (#497)
Resolves compile error with Parameters.cpp
Turns on AllParameters.cpp unit test but this is not resolved yet.
Changed dummy adapter to use BasicIdentifierAdapter.
Cuts several unused parameters.
Moved a few parameters.

This update is still in progress as there are some issues to resolve
with parameters properly failing for bad strings when they are
double or int types.