Tpetra::MultiVector: Fix #638 (potential DualView allocation bug)
@trilinos/tpetra

In the case where the zeroOut argument to Tpetra::MultiVector's
constructor is false, MultiVector may have been allocating its
DualView incorrectly.  Here are the offending lines of code (in the
internal allocDualView function):

  dual_view_type dv (d_view, h_view);
  dv.template modify<typename dual_view_type::t_dev::memory_space> ();
  // !!! BUG BELOW !!!
  // !!! This returns a _new_ DualView, not dv above !!!
  return dual_view_type (d_view, h_view);

The problem with this code, is that it it fails to mark the returned
DualView as modified on device.  Even though setting zeroOut = false
indicates that the user does not care about the MultiVector's initial
contents, it's still bad for the flags to be out of sync.

The reason this bug never manifested, is because Tpetra currently
requires enabling Kokkos' UVM support option.  This makes Cuda's
default memory space CudaUVMSpace, not CudaSpace.  As a result, both
the device and "host" View in a DualView are the same View, living in
CudaUVMSpace.  This means that even in CUDA builds, sync and modify in
Tpetra currently do nothing.

Build/Test Cases Summary
Enabled Packages: TpetraCore
Disabled Packages: FEI,PyTrilinos,Moertel,STK,SEACAS,ThreadPool,OptiPack,Rythmos,Intrepid,ROL,Panzer
0) MPI_DEBUG => Test case MPI_DEBUG was not run! => Does not affect push readiness! (-1.00 min)
1) SERIAL_RELEASE => Test case SERIAL_RELEASE was not run! => Does not affect push readiness! (-1.00 min)
2) MPI_DEBUG => passed: passed=107,notpassed=0 (36.00 min)
3) MPI_RELEASE => passed: passed=94,notpassed=0 (24.71 min)
Other local commits for this build/test group: 5d8efd9, d9b4ff8