Zoltan2: ParameterList updates with Teuchos changes (#612 #650)
This update includes:

A new BoolParameterEntryValidator - accepts true/false/"true"/"false"
A new Teuchos unit test for BoolParameterEntryValidator

Updates to AnyNumberParameterEntryValidator
  Added std::stod and std::stoi for C11 only.
  Added a Teuchos unit test to validate settings.

Also updated code in Zoltan2 to reflect these changes.

Changed the parameters function in Zoltan2 to be getValidParameters.
as discussed.

Note that this branch may likely be split up into several
since it has Teuchos and Zoltan2. However for discussion and testing
it's convenient to have this together for the moment.