Zoltan2: EnhancedNumberValidator given string support (#497)
Added string support for EnhancedNumberValidator.
So "5" and 5 will both be accepted for an int type.

This was done so all Zoltan2 parameters will work with strings.

EnhancedNumberValidator is templated to type T while
AnyNumberParameterEntryValidator explicitly works with
int, double, string.

This means the conditions for string conversion are different.
Currently EnhancedNumberValidator is checking the T type
explicitly to determine if the std::atoi or std::stoi
conversions (depending on CXX11) are appropriate.

If none match the default is a double conversion using
std::atof or std::stod (again depending on CXX11).

Probably we should consider refactoring the handling of
strings to the base class so that AnyNumberParameterEntryValidator
and EnhancedNumberValidator can share and unify this code better.