Jamfile: Add use of install-bin-dir for end-user installs.
This commit changes all Jamfiles to use the new install-bin-dir constant
for the directory into which the end-user executables are installed. This
allows the Jamroot file to determine what the bin directory needs to be.
By default it will be $(install-root)/bin, but in cases where a machine
has multiple architectures (knl vs haswell) the bin directory can be set
to something specific to that architecture.

Change-Id: I29769d39b75d14d57874238d59923a1cd7bc911e

(From STK/SEACAS)