Sacado:  Don't pass LayoutStride through createDynRankView().
If a user calls createDynRankView() with an input view with
LayoutStride, the resulting view also was getting LayoutStride.  This
doesn't really work, and is probably not what the user intended.  So
instead we now replace LayoutStride with the default layout for the
input view execution space.  Thanks to Kyungjoo for reporting this.