KokkosKernels: Add macros for whether to build for specific device types
Add macros of the form
TPETRAKERNELS_BUILD_EXECUTION_SPACE_${EXEC_SPACE} to
TpetraKernels_config.h.  These macros tell KokkosKernels whether to
build kernels for that Kokkos execution space.  This relates to #700,
in that it will help us reduce the number of kernels to build.  In
particular, we only want to build kernels for execution spaces that
Tpetra uses, not for all execution spaces that Kokkos enables.

NOTE: This changes CMake logic.  Please change bjam accordingly.