KokkosKernels: Start specializing SPMV_MV on whether Scalar is integer
SPMV_MV implements sparse matrix-vector multiply for multiple
right-hand sides.  It currently produces most of the code for sparse
matrix-vector multiply.  This commit adds a Boolean
'integerScalarType' template parameter to SPMV_MV.  This parameter
defaults to true if and only if the matrix's Scalar type is an
integer.  In subsequent commits, we will use this template parameter
to make SPMV_MV use a less heavily optimized implementation if the
matrix's Scalar type is an integer.  This will address #700.