Fix Apollo and Perseus nightly testing
The nightly testing on these machines stopped reporting after the
module change due to the modules not being available. The cause is
that Jenkins didn't have any of the modules available due to the
shell not properly sourcing the startup files. The solution is to
manually source the right scripts to make the SEMS environment
available.

There was also a missing load of a compiler similar to Artemis which
I fixed as well.