Xpetra: Fix guards for Kokkos interfaces
The test cannot run if Tpetra is disabled. In contrast to the Kokkos interface
in Xpetra::MultiVector the Kokkos interface in CrsMatrix requires Tpetra as
the underlying Kokkos objects Kokkos::CrsMatrix reside in Tpetra. That is, there
is no Kokkos interface if Tpetra is disabled. If both Epetra and Tpetra are enabled
the Kokkos interface is "emulated" for Epetra and will work. But it will not work in
an Epetra-only build

Remove explicit calls to Kokkos::initialize and Kokkos::finalize. See also issue 522