Teuchos: Move EReductionType & its fns into a new header file
@trilinos/teuchos

Teuchos::EReductionType used to live in the same header file as a huge
mass of wrappers for other MPI functionality.  It's nicer to be able
to include the one header file if that's all you need.  This commit
splits off EReductionType and its functions (conversion to MPI_Op, and
conversion to human-readable string) into their own header file.  It
also adds better documentation for EReductionType.