Stokhos:  Fix magnitudeType insanity for UQ::PCE scalar type.
For a long time, magnitudeType for a UQ::PCE scalar type was again a
UQ::PCE scalar type.  This was a workaround for issues in Teuchos::BLAS
where it is assumed that if the scalar type is not complex, the scalar
type is assignable to its magnitude type.  I have since figured out how
to workaround those issues with the MP::Vector scalar type when making
magnitudeType an actual magnitude.  This ports the workaround to the
UQ::PCE scalar type, so that its mangitude type is actually a magnitude.
This fixes an issue that came up recently in MueLu for comparisons
between magnitude types of different precisions.