Xpetra: move blocked classes from sup to src in Xpetra
In order to use blocked operators, multivectors and maps natively in MueLu
we move them from the supplementary stuff to the main src folder. We need them
there such that we can extend, e.g., the Xpetra::MultiVectorFactory to create
BlockedMultiVectors as well as standard MultiVectors without complicating the
dependencies and cmake settings.

Added a new class BlockedMap which is supposed to handle blocked maps (which
correspond to the ProductVectorSpace objects in Thyra). Most of the functionality
from MapExtractor will be moved to the BlockedMap class. Sooner or later we should
be able to remove MapExtractor as all functionality is in BlockedMap and the factory
classes (e.g. MultiVectorFactory)

Some cleanup of includes; use forward declarations to avoid circular file inclusions