Sacado:  Fix subdynrankview() for Sacado scalar types.
CDash showed a test failure for subdynrankview() for the new
LayoutContiguous view specialization, which exposed some bugs in how the
dimensions of the resulting subview were set (for both the original and
LayoutContiguous specializations).  In both cases, the errors "appeared"
to be harmless, in that the indexing was still correct even with the
bugs (but enabling bounds checking showed the offset mapping was not
consistent), so this shouldn't affect any user code.  I also added a few
more checks in the unit tests to make sure the dimensions of the
resulting subviews were as expected.