Add Trilinos options for float and complex off by default (#362)
Now there are top-level options for float and complex and these are off by
default as per the conversation and decision documented in
trilinos/Trilinos#362.  This makes it easy to enable and disable support for
these types.

Some work needs to be done to get some of the packages to allow enabling and
disabling float, std:complexfloat> and std::complex<double> somewhat
independently but that can be done later.  Really only Tpetra allows this
correctly.

Also, these are disabled by default to speed up the default build of Trilinos
and very few customers use or need float and std::complex types.  As a matter
of fact, float support in Tpetra and downstream is currently broken.  This is
likely due to the fact that Tpetra unconditionally disabled float by default
and it is doubtful that anyone has tested this in a long time (and there are
no automated tests that test float support).

I fixed the logic for Tpetra complex<double> support in Anasazi so that Tpetra
(take a look at the updated logic).

I also fixed some typos in documentation.