MueLu: use map of X for correction vector in Hierarchy::Iterate
The idea is, that the correction vector is of the same type then the X vector
(i.e. either MultiVector or BlockedMultiVector). The BlockedCrsMatrix::apply
routine can deal with all combinations of blocked and non-blocked multi vectors.
The MultiVector::update routine, though, only accepts multi vectors based on Map.
Only the BlockedMultiVector::update routine accepts multi vectors based on Map or
BlockedMap.

In Hierarchy::Iterate, the coarse rhs and solution vector may or may not live on
a BlockedMap (depending on the type of the operator). On the finest level, the fine
level B and X are of the type the user has provided (based on Map or BlockedMap)