PyTrilinos: Add AsMap() method to Epetra.BlockMap class
As per #772, the Epetra.CrsMatrix constructor requires an Epetra.Map,
and *not* an Epetra.BlockMap. If you attempt to obtain the Map from
Epetra.Vector.Map(), you will get a BlockMap, not a Map, and the
constructor will fail. By adding this AsMap() method, I allow the user
to attempt to cast a BlockMap to a Map.