Move MPI_RELEASE_DEBUG_SHARED_PT_COMPLEX out of --default-builds (#482)
The way that I was including MPI_RELEASE_DEBUG_SHARED_PT_COMPLEX in
--default-builds and then excluding it in checkin-test-sems.sh created messy
amended commit messages like:

    Build/Test Cases Summary
    Enabled Packages: Teuchos
    Disabled Packages: PyTrilinos,Claps,TriKota
    0) MPI_RELEASE_DEBUG_SHARED_PT => passed: passed=123,notpassed=0 (2.46 min)
    1) MPI_RELEASE_DEBUG_SHARED_PT_COMPLEX => Test case MPI_RELEASE_DEBUG_SHARED_PT_COMPLEX was not run! => Does not affect push readiness! (-1.00 min)

That is bad.

Therefore, for now, I have moved the definition of the
MPI_RELEASE_DEBUG_SHARED_PT_COMPLEX build to the checkin-test-sems.sh script
so that it can be run with --st-extra-builds.  This has the disadvantage that
it would allow changed ST packages that are not PT packages to be enabled
which would cause them to ben enabled in the inner configure and therefore
enable more that PT packages.  However, given the small number of remaining
number of ST packages that are not also PT packages, this is a msll risk.
Also, many of the ST but not PT package are terminal packages so enabling them
would not change the confiugre and build of upstream PT packages.

Build/Test Cases Summary
Enabled Packages:
Disabled Packages: PyTrilinos,Claps,TriKota
Enabled all Packages
0) MPI_RELEASE_DEBUG_SHARED_PT => passed: passed=2260,notpassed=0 (104.50 min)