Set SEMS ParMETIS libs explicitly  (#945)
For some versions of CMake (2.8.12 and 3.7.0) and GCC (4.7.2), CMake will
refuse to find the SEMS-installed ParMETIS libs with FIND_LIBRARY().  I have
not seen this problem with any other TPL other than the SEMS-installed version
of ParMETIS.

I have pinged Kitware about this issue and am working to get Brad King access
to our env so that he can reproduce this issue.  But until this can be fixed,
this is a workaround to allow us to use newer CMake versions with the SEMS
env.  This should allow the usage of the checkin-test-sems.sh script on RHEL 7
systems with the SEMS env (see #945).

Build/Test Cases Summary
Enabled Packages:
Disabled Packages: PyTrilinos,Claps,TriKota
Enabled all Packages
0) MPI_RELEASE_DEBUG_SHARED_PT => passed: passed=2259,notpassed=0 (87.86 min)