Stratimikos: allow compilation without Epetra
allow an optional, not required, dependency
on Epetra. it seems only the builder header
needs #ifdef guards.
since this changes the dependency graph,
it is protected by

Stratimikos_MAKE_EPETRA_OPTIONAL:BOOL

which must explicitly be set to true
for the changes in dependency status
to change.
this flag should remain until the next
backwards-compatibility-breaking
release of Trilinos.