Fix potential issue with solver when matrix is a view.
Issue #1035:  There is potential to get the wrong behavior in the
SerialDenseSolver class when iterative refinement is being used.
A copy of the original matrix is kept separate from the factored
matrix.  However, the copy constructor could just copy the view
if the original matrix is a view of a larger matrix.  Adding the
explicit Teuchos::Copy to the copy constructor will ensure the
correct behavior.