Tpetra::RowGraph: Add "row view" methods
@trilinos/tpetra CrsGraph has supportsRowViews(), getGlobalRowView(),
and getLocalRowView(), analogous to CrsMatrix.  However, RowGraph (the
parent class of CrsGraph) did _not_ have these methods, which broke the
analogy with RowMatrix.  This commit fixes that.

The point of this commit is to factor the makeColMap() method out of
CrsGraph, into a separate method that CrsGraph can use.  The resulting
makeColMap() implementation cannot include Tpetra_CrsGraph.hpp, since
that would otherwise introduce a circular dependency.  makeColMap()
really has no need of CrsGraph's entire interface; it just needs access
to the entries in sequence.