Removing any documentation, discussion or hint about scaling operators in Belos.
Since the beginning of time (which was around 2003 :) ) there has been member variables
for left and right scaling in the Belos::LinearProblem.  This was inspired by the
Epetra_LinearProblem, yet never completely implemented in Belos.  To do so would require
the extension of the MultiVector traits class and it is not clear that this is necessary.
So, to avoid confusion, any documentation about scaling in the Belos::LinearProblem is
being removed.