MueLu: finish Kokkosification of unamalgamation routines in TentativePFactory
We do the unamalgamation directly in the TentativePFactory using the information
from the AmalgamationInfo container class. We do not call the helper functions as
we would need a separate _kokkos class for the AmalgamataionFactory then (which we
want to avoid).
In general, it turns out to be a bad idea to have routines working on data within
container classes. We should keep algorithms and data strictly separate.