Teuchos: Skip return only for NVCC (#1133, #1140)
After talking with Roger P. I have changed this to only remove the return
statement for __NVCC__.  This is because only __NVCC__ warns about the
unreachable return but having that return statement leads to safer code by
avoiding undefined behavior in case a function is modified in such a way that
this return is no longer unreachable.  In that case, we want it to return
something to avoid undefined behavior.

I also improved the documentation some.