Tempus: Ensure that the order gets set on the SolutionState.
Some of the steppers were not setting the order on the state
after each time step.  This should be done in case,
 * the stepper changes the order from one time step to the next, or
 * the stepper gets a SolutionState from another stepper with a
   different order.
Also changed the order on the initial state (i.e., initial
condition) from the minimum order to the current stepper order.

This should fix an error in the reported order noted by I. Tezuar.