PyTrilinos: Check for doxygen 1.8.13
Doxygen version 1.8.13 has a bug that results in a segmentation fault
when run against certain PyTrilinos doxyfiles. The build system now
generates a fatal error if doxygen is version 1.8.13 and PyTrilinos
docstrings are turned on. To get around this fatal error, the user
must turn off PyTrilinos docstrings or choose a different version of
doxygen. An error message is printed to this effect.