Sacado:  Add initial work on new design for Fad expression templates.
This commit adds a new implementation of the standard Fad classes (DFad,
SFad, SLFad, ViewFad) that uses a more modern expression template
design.  It also uses several C++11 features, so it is only usable with
C++11.  The classes are in the Sacado::Fad::Exp namespace, and should be
usable by any code just by changing the Fad type.  I also added a CMake
option:

-D Sacado_NEW_FAD_DESIGN_IS_DEFAULT:BOOL=ON \

which replaces the original Fad classes with the new ones.  This is
primarily for testing purposes so that I can test it works without
duplicating all of the unit tests.  I have not yet explored peformance
of the new approach, but one of the goals of this new design is to allow
easier specialization of the expression templates, primarily for nested
AD/ensemble types.  It also unifies the implementation of SFad with the
others.