Adding conditionals, based on TEUCHOSCORE_CXX11, to address Epetra_DataAccess scoping
To address C++11-compliant compilers that now require namespace scoping
of the Epetra_DataAccess enumeration, several Epetra interface changes
were made in Ifpack and Belos.  Unfortunately this is breaking the
Trilinos-Xyce build that tests non-C++11-compliant compilers, issue #1145.
So the use of scoping on the Epetra_DataAccess enumeration is being
conditionally compiled based on the TEUCHOSCORE_CXX11 flag for Ifpack
and Belos, which are required the build cleanly with non-C++11-compliant
compilers.