Teuchos: fix the default values for help option for CommandLineProcessor
Depending on whether help option was before or after some arguments on
the command line, it could affect the output. Specifically, the provided
command line values of other parameters could appear as the default ones
in the help output.

This patch fixes (or works around) the problem by first searching
through the parameters to find if help option is present.

Fix #585.