Tpetra: Fix the generalization of #1345
@trilinos/tpetra Whenever doing Kokkos::parallel_* in Tpetra, always
use Kokkos::RangePolicy with an explicit execution space.  Do not use
an integer for the range argument.  Integer ranges make Kokkos use the
default execution space, which may not be the right one for the Tpetra
object or function.

This commit changes one explicit use of an integer range that worked
around a build error with CUDA builds last year.  That build error
doesn't exist any more, so I just switched that code to use a
Kokkos::RangePolicy instead of an integer range.

This commit does _not_ address the experimental sparse matrix-matrix
multiply capability.  I will deal with that in my next commit.  I
wanted to keep that separate, since it does not build by default.