PyTrilinos: Fixed memory leak in specific Teuchos::RCP typemaps
For a variable type VAR that is stored internally as a Teuchos::RCP,
the "directorin" typemaps for const and non-const versions of VAR* and
VAR& arguments had a memory leak fixed here.

This cropped up, for example, for the Epetra.Operator class, where the
user would inherit from Epetra.Operator and define an Apply(X,Y)
method. In C++, X and Y are const and non-const Epetra_MultiVector&,
respectively. In Python, they need to be converted to Teuchos::RCP<
(const) Epetra_MultiVector > before being wrapped, and this conversion
had a memory leak.