MueLu: in Hierarchy::Iterate(), cache the result of description() and ShortClassName(), and use a new Monitor constructor that accepts strings instead of calling these.
Details:
description() calls GetGlobalNumLevels() and GetOperatorComplexity(), both of which have nontrivial cost.  GetGlobalNumLevels() in particular performs an MPI_Reduce.  In cases where there are many calls to Iterate() for a relatively small matrix, these calls have a measurable cost as a percentage of the total runtime.

Similarly, ShortClassName() results in a call to Teuchos::demangleName(), which as previously noted (in issue #1432) has measurable cost.