Anasazi: add an early warning about small subspace dimensions
If a user specifies a "Maximum Subspace Dimension" to be too small, and
then during the solution the solver restarts, it will fails with a
"cryptic" message

    Throw test that evaluated to true: state.curDim <= d_restartDim

This does not actually tell a user what exactly is happening, and how to
address the problem. This patch tries to address it.

Build/Test Cases Summary
Enabled Packages: Anasazi
0) MPI_RELEASE_DEBUG_SHARED_PT => passed: passed=71,notpassed=0 (5.98 min)