Add a new prototype for continuous testing
As part of moving the continuous testing off of sadl30906 it was decided
to also improve the ability to update the compiler easily and to be able
to run more than one continuous build off a single configuration. With the
success of the parameterized nightly builds the same methods for making
those builds configurable through jenkins was applied to the continuous to
see if the continuous could be made as flexible.