MueLu:  Fixed an annoying ETI build error
We're now in a situation that we need a near-sorcerous dance in
every single (non-unit) test to handle weird instantiation types.

Of course, the dance is (needlessly) different in each file.

We should really abstract that out somewhere in TestHelpers and then
call that.

Consider this a band-aid at present.

Build/Test Cases Summary
Enabled Packages: MueLu
Disabled Packages: PyTrilinos,Claps,TriKota
0) MPI_RELEASE_DEBUG_SHARED_PT => passed: passed=53,notpassed=0 (14.92 min)