Tpetra::CrsGraph: Sep owned vs. nonowned rows in filtered global insert
@trilinos/tpetra In Tpetra::CrsGraph, separate the implementation of
the owned row case vs. the nonowned row case, in filtered global
insert (insertGlobalIndicesFiltered).  This will help with thread
parallelization of unpackAndCombine, and in general with reducing
statefulness.

Column filtering only exists because that's what Epetra did.  I've
rarely seen a user like it, because it tended to hide their bugs until
they manifested in solvers, in hard-to-find ways.  It's supposed to
have been useful for things like LocalFilter, even though that uses a
different implementation.