Zoltan2: Fixes variable size handling
This removes some unnecessary exchanges and makes
the directory work with variable sizes in a manner which
I think is much closer to the original intention.

The communicator do_wait function has been modified to
handle the variable size. That part will need further
discussion. I have also modified the resize method to
handle the case which was previously marked as unsupported.

Even though this is working for Add and Replace with
variable length arrays, I'm not confident this is the
best way to organize things and will need further
discussion.

Note searching for NEW METHOD will show the 5 places I think
I need to discuss.

This commit  fixes a bunch of other issues with the directory.