Tpetra: Move transformLocalValues impl out of CrsGraph into CrsMatrix
@trilinos/tpetra Move the implementation of
CrsMatrix::transformLocalValues entirely out of CrsGraph, into
CrsMatrix as a nontemplated method.  This reduces the amount of
implementation code in Tpetra_CrsGraph_decl.hpp, and should thus
improve build time.

Internally, use std::function to represent the transformation function
f, instead of a template parameter.  This may have a slight run-time
overhead, but transformLocalValues is a rarely used public method, so
it's worth the benefit of hiding the implementation.