Tpetra,Stokhos: Fix #1658 and #1088
@trilinos/tpetra @trilinos/stokhos It's unfortunate that I had to
merge these two fixes into one commit.  The reason is that #1658
blocks #1088, but I couldn't reorder the two commits without the risk
of losing information.  I will give the two commits' messages below.

* Stokhos,Tpetra: Fix #1658 (Tpetra::MultiVector unpack kernels)

@trilinos/tpetra Tpetra::MultiVector's unpack kernels (used in
unpackAndCombineNew) now take a Kokkos execution space argument.
Before, the kernels were using the output Kokkos::View (the
MultiVector's data) to determine the execution space.  The problem
with this, is that for a Kokkos::DualView of CudaUVMSpace, the host
and device Views are the same.  This blocks #1088, whose fix requires
that the input View (the buffer to unpack) be either a CudaSpace or a
HostSpace View.  (CudaUVMSpace::execution_space == Cuda, so the kernel
will always run on device, even given a HostSpace buffer to unpack.)

The new execution space input argument lets the user specify the
execution space on which to run.  It's an execution space instance, so
this could give us a future option to run in a separate CUDA stream
(e.g., for overlap of communication and computation), or to run on a
subset of threads.

I also fixed the unpack kernel specializations in Stokhos.

* Tpetra: Fix #1088 (never give CUDA UVM allocations to MPI)

@trilinos/tpetra This commit does two things:

  1. Change DistObject so that it never gives CUDA UVM (i.e.,
     Kokkos::CudaUVMSpace) allocations to MPI sends or receives.

  2. Make packCrsMatrix and unpackCrsMatrix use ETI (explicit template
     instantiation), to reduce build time and speed up debugging of
     the pack and unpack routines.

I did the first of these by insisting on Kokkos::CudaSpace, rather
than Kokkos::CudaUVMSpace, for all of DistObject's communication
buffers: imports_, numImportPacketsPerLID_, exports_, and
numExportPacketsPerLID_.  I added a public typedef to DistObject,
buffer_device_type, to help users allocate communication buffers of
the right type.  This required changing Tpetra::MultiVector and
Tpetra::Details::CooMatrix, which are currently the only two
implementations of the "new" DistObject interface.

This commit fixes #1088.  However, this commit will break CUDA tests
if the MPI implementation is not CUDA aware.  See #1088 and #1571.

I also added a lot of optional debug output.  For example,
doTransferNew's debug output now says whether the communication
buffers live on host or device.  The debug output is off by default.
If you would like to enable it temporarily for debugging, just search
for the string "constexpr bool debug = false"; the comment above this
line explains what to do.

The second thing I did was to make packCrsMatrix and unpackCrsMatrix
use explicit template instantiation (ETI).  This reduces build time.
I did it specifically to speed up debugging of my changes to the pack
and unpack routines.  [Author's note: This is one reason why it was
hard to reorder the two commits.]

Stokhos' ETI system is more restrictive than Tpetra's ETI system, so I
had to rename unpackCrsMatrix's headers.  Sorry to folks who are
working on this at the moment.

I also fixed possible errors in Tpetra::CrsMatrix::*[gG]aussSeidel*
just to make sure.