Zoltan2: Updates for directory mechanics
Testing infrastructure developed to be more thorough.
The unit test now grinds through many combinations of
total Ids, vector/single user_t, different gid_t,
gid_t as a single value or set of values, and also
tests Add, Replace, and Aggregate.

Note that only Kokkos mode tests all of Add, Replace, Aggregate
along with vector user_t. Other modes are for development
and eventually will go away.

Added a simple test directoryTest_KokkosSimple for easier example of
using the directory. The main test is thorough but getting
opaque so this is intended as a simple example of how to
setup the directory and use it.

Added testing for when gid is a set of values such as [1,2,8,10].
The multple gid type must be set up as some kind of struct/class.
For example: struct gids { int x[4]; };

It should work with a default constructor gids().
size_of(gids) should be valid (it can't contain mem allocs)
It should be copyable/comparable, so a=b works if a and b are both gids type:
Otherwise the content of the struct/class is arbitrary.

This update removes the #ifdef macros which distinguished between
the vector user_t mode and the single type mode. Now they can both
work simultaneously and both are setup and tested in the
unit test. Presently I am using two derived Directory classes to
handle with/without variable vector user_t. Not clear if this will be the
best API setup.